refactor: Chat 컴포넌트 계층적으로 폴더 구조 변경
