+++ client/src/components/Chat/ChatList/ChatItem/ThreadPreview/style.ts
