+++ client/src/components/Chat/ChatList/ChatItem/style.ts
